Now using TTile for tiles, also other minor edits to comments and code

+++ Graal Utilities/TGmap/TGmap.cpp
@@ -31,8 +31,7 @@ bool TGmap::load(const char* filePath) {
     ifstream myfile (filePath);
     
     //check to see if the file is opened:
-    if (myfile.is_open())
-    {
+    if (myfile.is_open()) {
         // Check header
         getline (myfile,line);
         if ( line != "GRMAP001" ) return 0;
@@ -42,8 +41,7 @@ bool TGmap::load(const char* filePath) {
         
         // while there are still lines in the
         // file, keep reading:
-        while (! myfile.eof() )
-        {
+        while (! myfile.eof() ) {
             // place the line from myfile into the
             // line variable:
             getline (myfile, line);
@@ -52,13 +50,14 @@ bool TGmap::load(const char* filePath) {
             if ( line.substr(0, 5) == "WIDTH" )   this->width   = atoi(line.substr(5, -1).c_str());
             if ( line.substr(0, 10) == "MINIMAPIMG" ) this->minimap = line.substr(11, -1).c_str();
             
+            // this is where we stop reading the level names
             if ( line.substr(0, 13) == "LEVELNAMESEND" ) mode = 0;
             
             std::vector<string> test(Tokenize(line, ','));
             
             // We need to know that were reading levels, not just that mode == 1, this needs updated
             if ( mode == 1 && line != "" ) {
-                for ( int i = 0; i < test.size(); i++ ) {
+                for ( size_t i = 0; i < test.size(); i++ ) {
                     TLevel newLevel;
                     newLevel.name = test[i].substr(1,test[i].length()-2);
                     newLevel.loadLevel(extractPath(filePath) + newLevel.name);
@@ -66,6 +65,7 @@ bool TGmap::load(const char* filePath) {
                 }
             }
             
+            // This is where we start reading the level names
             if ( line.substr(0, 10) == "LEVELNAMES" ) mode = 1;
             
         }
